// src/models/_foodb-models/db.d.ts
export interface compoundsHealthEffectsAttribute {
	id: number;
	compoundId: number;
	healthEffectId: number;
	origHealthEffectName?: string;
	origCompoundName?: string;
	origCitation?: string;
	citation: string;
	citationType: string;
	createdAt?: Date;
	updatedAt?: Date;
	creatorId?: number;
	updaterId?: number;
	sourceId?: number;
	sourceType?: string;
}

export interface compoundsHealthEffectsInstance extends Sequelize.Instance<compoundsHealthEffectsAttribute>, compoundsHealthEffectsAttribute {
}